# Experiments with parallelization of the Collins-Krandick algorithm