Algebraic numbers aren't pickable, that should be fixed.

Some benchmarks.

With
```python
In [1]: %%time
      : e = x**40 - 15*x**17 - 21*x**3 + 11
      : 
      : R = ZZ.poly_ring(x)
      : p = R.from_expr(e)
      : 
      : complexes = R.dup_isolate_complex_roots_sqf(p, blackbox=True)
      : 
      : 
CPU times: user 7.04 s, sys: 141 ms, total: 7.18 s
Wall time: 1min 43s
```

Without:
```python
In [1]: %%time
      : e = x**40 - 15*x**17 - 21*x**3 + 11
      : 
      : R = ZZ.poly_ring(x)
      : p = R.from_expr(e)
      : 
      : complexes = R.dup_isolate_complex_roots_sqf(p, blackbox=True)
      : 
      : 
CPU times: user 2min 51s, sys: 279 ms, total: 2min 51s
Wall time: 2min 51s
```

PS: for https://github.com/diofant/diofant/issues/823