@skirpichev 2019-03-31
depends on pickling support, e.g. #826