@skirpichev 2020-03-07
New benchmark.

On 565a433356:
```
In [1]: %%time
      : e = x**40 - 15*x**17 - 21*x**3 + 11
      : R = ZZ.poly_ring(x)
      : p = R.from_expr(e)
      : complexes = R.dup_isolate_complex_roots_sqf(p, blackbox=True)
      : 
      : 
CPU times: user 21.1 s, sys: 28.8 ms, total: 21.2 s
Wall time: 21.4 s
```

On the branch:
```
In [1]: %%time
      : e = x**40 - 15*x**17 - 21*x**3 + 11
      : R = ZZ.poly_ring(x)
      : p = R.from_expr(e)
      : complexes = R.dup_isolate_complex_roots_sqf(p, blackbox=True)
      : 
      : 
CPU times: user 2.91 s, sys: 36.6 ms, total: 2.95 s
Wall time: 13.4 s
```